Provide symbols files in all library packages
Based on the autogenerated files for i386 available from
http://qa.debian.org/cgi-bin/mole/seedsymbols.

The exported symbols are identical for the 32-bit and 64-bit
architectures, so we can just use the #PACKAGE# trick, symlink the
lib{32,64}*.symbols files to their counterparts and be done with it.
This way we only need to maintain two symbols files instead of five.