Add cons25-debian terminfo entry to ncurses-base
The Debian GNU/kfreeBSD console differs from the stock FreeBSD one in
respect to the backspace/delete keys, which have been modified to
behave as in Linux to match the Debian Policy.  This requires a new
terminfo entry because the backspace key shows erratic behavior with
the standard cons25 entry.  Closes: #607662