Omit -L part from ncurses5{w,}-config output
"ncurses5-config --libs" includes a "-L" part which is not needed
since the library is installed in a directory in the default linker
search path.  The value is also wrong if ncurses-bin is of a
different architecture than libncurses5-dev.

Upstream included a similar patch in the 20110507 patchlevel to omit
the -L part if $libdir is /usr/lib.  Until there is a way to obtain
the multiarch path without running dpkg-architecture our patch is not
upstreamable.