Fix FTBFS when not building from git
Commit 0b403f0b648ecc0d64e106e83df36da68fabb029 has the problem that
dpkg-source by default "helpfully" omits debian/libtermcap.so when
creating the debian.tar.gz.  Rather than copying the whole list of
ignored-by-default file patterns from "dpkg-source --help" output,
just ignore .gitignore since this is the only file that we should not
include and which is not deleted by dh_clean.

It would be nice if dpkg-source had an "unignore" option in we could
specify the .so pattern, see http://bugs.debian.org/615828.