Replace various /usr/share/doc directories with symlinks
This saves about 180 Kilobyte for each package where this is done,
bringing down the maximum number of copies of the big upstream
changelog from twenty to six.

Add necessary postinst scripts to convert directories to symlinks on
upgrade.  Until dpkg-maintscript-helper or debhelper provide
interfaces for the directory → symlink conversion, this needs to be
done manually.  Everybody seems to do that slightly differently, the
snippets used here are based on Colin Watson's libpipeline-dev
package.