Assume working poll() in biarch builds
The corresponding configure check cannot be run when cross-compiling,
due to that problem the biarch packages were still affected by
http://bugs.debian.org/676461 (default is cf_cv_working_poll=unknown
after which select() will be used rather than poll()).

Also set cf_cv_type_of_bool to unsigned char in the 32-bit biarch
builds like it had been done already for the 64-bit one.  This is not
strictly necessary since it defaults to that anyway, but 'configure'
issues a warning.