Drop the dependency of the biarch packages on libtinfo5
The lib32/lib64 packages should be dropped at some time in the future
in favor of multiarch, but users may want to keep them a little
longer, e.g. because third-party software depends on these packages.
They would be forced to hold back the native libraries and the
ncurses-bin package as well, which is not good.

The dropped dependency also means that we have to convert
/usr/share/doc/lib{32,64}tinfo5 back to directories and adjust the
symlinks in the other biarch packages.