Correct /usr/share/doc symlinks on upgrades
Unfortunately this is necessary, because dpkg does not always do it.
For instance, if lib32ncurses5 is unpacked before lib32tinfo5, dpkg
notices that /usr/share/doc/lib32ncurses5 is a symlink to the same
directory in both the old and the new version of the package and does
not update the symlink.